requirements.txt
# Core data manipulation
pandas>=2.0.0
numpy>=1.24.0

# Data sources
yfinance>=0.2.0
fredapi>=0.5.0
requests>=2.31.0

# Visualization
plotly>=5.18.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web dashboard
dash>=2.14.0
dash-bootstrap-components>=1.5.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.1.0
notebook>=7.0.0

# Data caching and storage
python-dotenv>=1.0.0
sqlalchemy>=2.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
